src/neahtta/requirements.txt
Babel==0.9.6
Flask==0.9
Flask-Actions==0.6.5
Flask-Babel==0.8
Jinja2==2.6
PyYAML==3.10
Werkzeug==0.8.3
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
gunicorn==0.17.2
lxml==2.3.3
pexpect==2.4
pytz==2012j
requests==1.1.0
speaklater==1.3
wsgiref==0.1.2
